removed option A benefit filter
