+++ __tests__/selectors/benefits_test.js
@@ -103,25 +103,19 @@ describe("getFilteredBenefits", () => {
     };
   });
 
-  // don't show benefit 0 because it's not independent
-  it("displays benefits 1, 2 and 3 if nothing selected", () => {
+  it("displays all benefits if nothing selected", () => {
     let returnValue = getFilteredBenefits(state, props).map(b => b.id);
     returnValue.sort();
-    expect(returnValue).toEqual(["1", "2", "3"]);
+    expect(returnValue).toEqual(["0", "1", "2", "3", "4"]);
   });
 
-  // 0 and 2 match, but 0 is a child of 2, so we just show 2
-  it("display benefits 2 if patronType p1", () => {
+  it("displays benefits 0, 2, 4 if patronType p1", () => {
     state.patronType = "p1";
-    expect(getFilteredBenefits(state, props).map(b => b.id)).toEqual(["2"]);
-  });
-
-  // eligible for 1, 3, 4. Need only matches 4, a child of 3. So show 3
-  it("shows an eligible parent for a matching child", () => {
-    state.patronType = "p3";
-    state.needs = [{ id: "n0", benefits: ["4"] }];
-    state.selectedNeeds = { n0: "n0" };
-    expect(getFilteredBenefits(state, props).map(b => b.id)).toEqual(["3"]);
+    expect(getFilteredBenefits(state, props).map(b => b.id)).toEqual([
+      "0",
+      "2",
+      "4"
+    ]);
   });
 
   it("runs a lunr search on the english index if searchString is set an english is used", () => {
